Betting UI
Added wager entry and listview to log actions taken by user. One problem with the log so far is when switching to the spin wheel it erases the contents of the log. I believe the issue lies in using different controllers or fxml scenes, it creates new objects.
Also made AccountsManager a singleton. Was thinking about making the log a singleton but I don't think it would communicate with the controller