added exit confirm box
using either the x button or using the menubar, there is an exit confirm box displayed. if yes then saves player before exiting. otherwise goes back to game. there is a nullpointer if trying to exit on loginscreen because there is no player data loaded that can be saved. maybe fixed by detecting nullpointer or what scene user is on